consolidate cm.bufferwin~
