--- cm.bufferwin~/cmutil.h
@@ -1,33 +1,0 @@
-/*
- Circuit Music Labs utilities library file.
- Copyright (C) 2014  Matthias Müller - Circuit Music Labs
- 
- This program is free software: you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- (at your option) any later version.
- 
- This program is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU General Public License for more details.
- 
- You should have received a copy of the GNU General Public License
- along with this program.  If not, see <http://www.gnu.org/licenses/>.
- 
- circuit.music.labs@gmail.com
- 
- */
-
-#ifndef _cmutil_h
-#define _cmutil_h
-
-#define RANDMAX 10000
-
-// RANDOM NUMBER GENERATOR
-double cm_random(double *min, double *max);
-
-// LINEAR INTERPOLATION FUNCTION
-double cm_lininterp(double distance, float *b_sample, t_atom_long b_channelcount, short channel);
-
-#endif
